add --dry-run

+++ pkg/cmdimpl/add.go
@@ -44,6 +44,7 @@ type AddParamSet struct {
 	PrivateKey     string
 	DeploymentType string
 	Namespace      string
+	DryRun         bool
 }
 
 var (
@@ -53,7 +54,7 @@ var (
 // checkError will print a message to stderr and exit
 func checkError(msg string, err interface{}) {
 	if err != nil {
-		fmt.Fprintf(shims.Stderr(), "%s: %v\n", msg, err)
+		fmt.Fprintf(shims.Stdout(), "%s: %v\n", msg, err)
 		shims.Exit(1)
 	}
 }
@@ -98,9 +99,9 @@ func generateWegoSourceManifest() []byte {
 	checkAddError(err)
 
 	cmd := fmt.Sprintf(`create secret git "wego" \
-		--url="ssh://git@github.com/%s/%s" \
-		--private-key-file="%s" \
-		--namespace=%s`,
+        --url="ssh://git@github.com/%s/%s" \
+        --private-key-file="%s" \
+        --namespace=%s`,
 		getOwner(),
 		fluxRepoName,
 		params.PrivateKey,
@@ -110,12 +111,12 @@ func generateWegoSourceManifest() []byte {
 	checkAddError(err)
 
 	cmd = fmt.Sprintf(`create source git "wego" \
-		--url="ssh://git@github.com/%s/%s" \
-		--branch="%s" \
-		--secret-ref="wego" \
-		--interval=30s \
-		--export \
-		--namespace=%s `,
+        --url="ssh://git@github.com/%s/%s" \
+        --branch="%s" \
+        --secret-ref="wego" \
+        --interval=30s \
+        --export \
+        --namespace=%s `,
 		getOwner(),
 		fluxRepoName,
 		params.Branch,
@@ -128,13 +129,13 @@ func generateWegoSourceManifest() []byte {
 
 func generateWegoKustomizeManifest() []byte {
 	cmd := fmt.Sprintf(`create kustomization "wego" \
-		--path="./" \
-		--source="wego" \
-		--prune=true \
-		--validation=client \
-		--interval=5m \
-		--export \
-		--namespace=%s`,
+        --path="./" \
+        --source="wego" \
+        --prune=true \
+        --validation=client \
+        --interval=5m \
+        --export \
+        --namespace=%s`,
 		params.Namespace)
 	kustomizeManifest, err := fluxops.CallFlux(cmd)
 	checkAddError(err)
@@ -145,9 +146,9 @@ func generateSourceManifest() []byte {
 	secretName := params.Name
 
 	cmd := fmt.Sprintf(`create secret git "%s" \
-			--url="%s" \
-			--private-key-file="%s" \
-			--namespace=%s`,
+            --url="%s" \
+            --private-key-file="%s" \
+            --namespace=%s`,
 		secretName,
 		params.Url,
 		params.PrivateKey,
@@ -156,12 +157,12 @@ func generateSourceManifest() []byte {
 	checkAddError(err)
 
 	cmd = fmt.Sprintf(`create source git "%s" \
-			--url="%s" \
-			--branch="%s" \
-			--secret-ref="%s" \
-			--interval=30s \
-			--export \
-			--namespace=%s `,
+            --url="%s" \
+            --branch="%s" \
+            --secret-ref="%s" \
+            --interval=30s \
+            --export \
+            --namespace=%s `,
 		params.Name,
 		params.Url,
 		params.Branch,
@@ -176,13 +177,13 @@ func generateSourceManifest() []byte {
 func generateKustomizeManifest() []byte {
 
 	cmd := fmt.Sprintf(`create kustomization "%s" \
-				--path="%s" \
-				--source="%s" \
-				--prune=true \
-				--validation=client \
-				--interval=5m \
-				--export \
-				--namespace=%s`,
+                --path="%s" \
+                --source="%s" \
+                --prune=true \
+                --validation=client \
+                --interval=5m \
+                --export \
+                --namespace=%s`,
 		params.Name,
 		params.Path,
 		params.Name,
@@ -222,10 +223,20 @@ func getOwnerFromUrl(url string) string {
 }
 
 func commitAndPush(files ...string) {
+	if params.DryRun {
+		fmt.Fprintf(shims.Stdout(), `git pull --rebase && \
+                git add %s && \
+                git commit -m 'Save %s' && \
+                git push\n`,
+			strings.Join(files, " "),
+			strings.Join(files, ", "))
+		return
+	}
+
 	cmd := fmt.Sprintf(`git pull --rebase && \
-				git add %s && \
-				git commit -m 'Save %s' && \
-				git push`,
+                git add %s && \
+                git commit -m 'Save %s' && \
+                git push`,
 		strings.Join(files, " "),
 		strings.Join(files, ", "))
 	_, err := utils.CallCommand(cmd)
@@ -259,14 +270,22 @@ func Add(args []string, allParams AddParamSet) {
 	reposDir := filepath.Join(os.Getenv("HOME"), ".wego", "repositories")
 	fluxRepo := filepath.Join(reposDir, fluxRepoName)
 	appSubdir := filepath.Join(fluxRepo, "apps", params.Name)
-	checkAddError(os.MkdirAll(appSubdir, 0755))
+
+	if !params.DryRun {
+		checkAddError(os.MkdirAll(appSubdir, 0755))
+	}
 
 	owner := getOwner()
-	checkAddError(os.Chdir(fluxRepo))
+
+	if !params.DryRun {
+		checkAddError(os.Chdir(fluxRepo))
+	}
 
 	if err := utils.CallCommandForEffect(fmt.Sprintf("git ls-remote ssh://git@github.com/%s/%s.git", owner, fluxRepoName)); err != nil {
 		fmt.Printf("wego repo does not exist, it will be created...\n")
-		checkAddError(utils.CallCommandForEffectWithDebug("git init"))
+		if !params.DryRun {
+			checkAddError(utils.CallCommandForEffectWithDebug("git init"))
+		}
 
 		c, err := cgitprovider.GithubProvider()
 		checkAddError(err)
@@ -280,26 +299,36 @@ func Add(args []string, allParams AddParamSet) {
 			LicenseTemplate: gitprovider.LicenseTemplateVar(gitprovider.LicenseTemplateApache2),
 		}
 
-		err = cgitprovider.CreateOrgRepository(c, orgRef, repoInfo, repoCreateOpts)
-		checkAddError(err)
-
-		cmd := fmt.Sprintf(`git remote add origin %s && \
-			git pull --rebase origin main && \
-			git checkout main && \
-			git push --set-upstream origin main`,
-			orgRef.String())
-		checkAddError(utils.CallCommandForEffectWithDebug(cmd))
-	} else {
-		cmd := "git branch --set-upstream-to=origin/main main"
-		checkAddError(utils.CallCommandForEffectWithDebug(cmd))
+		if !params.DryRun {
+			checkAddError(shims.CreateOrgRepository(c, orgRef, repoInfo, repoCreateOpts))
+			cmd := fmt.Sprintf(`git remote add origin %s && \
+            git pull --rebase origin main && \
+            git checkout main && \
+            git push --set-upstream origin main`,
+				orgRef.String())
+			checkAddError(utils.CallCommandForEffectWithDebug(cmd))
+		} else {
+			fmt.Fprintf(shims.Stdout(),
+				`git remote add origin %s && \
+            git pull --rebase origin main && \
+            git checkout main && \
+            git push --set-upstream origin main`,
+				orgRef.String())
+		}
+	} else if !params.DryRun {
+		checkAddError(utils.CallCommandForEffectWithDebug("git branch --set-upstream-to=origin/main main"))
 	}
 
 	// Install Source and Kustomize controllers, and CRD for application (may already be present)
 	wegoSource := generateWegoSourceManifest()
 	wegoKust := generateWegoKustomizeManifest()
-	kubectlApply := fmt.Sprintf("kubectl apply --namespace=%s -f -", params.Namespace)
-	checkAddError(utils.CallCommandForEffectWithInputPipe(kubectlApply, string(wegoSource)))
-	checkAddError(utils.CallCommandForEffectWithInputPipe(kubectlApply, string(wegoKust)))
+	if !params.DryRun {
+		kubectlApply := fmt.Sprintf("kubectl apply --namespace=%s -f -", params.Namespace)
+		checkAddError(utils.CallCommandForEffectWithInputPipe(kubectlApply, string(wegoSource)))
+		checkAddError(utils.CallCommandForEffectWithInputPipe(kubectlApply, string(wegoKust)))
+	} else {
+		fmt.Fprintf(shims.Stdout(), "Applying wego platform resources...")
+	}
 
 	// Create app.yaml
 	t, err := template.New("appYaml").Parse(appYamlTemplate)
@@ -327,11 +356,15 @@ func Add(args []string, allParams AddParamSet) {
 	manifestsName := filepath.Join(appSubdir, fmt.Sprintf("%s-%s.yaml", params.DeploymentType, params.Name))
 	appYamlName := filepath.Join(appSubdir, "app.yaml")
 
-	checkAddError(ioutil.WriteFile(sourceName, source, 0644))
-	checkAddError(ioutil.WriteFile(manifestsName, appManifests, 0644))
-	checkAddError(ioutil.WriteFile(appYamlName, populated.Bytes(), 0644))
+	if !params.DryRun {
+		checkAddError(ioutil.WriteFile(sourceName, source, 0644))
+		checkAddError(ioutil.WriteFile(manifestsName, appManifests, 0644))
+		checkAddError(ioutil.WriteFile(appYamlName, populated.Bytes(), 0644))
 
-	commitAndPush(sourceName, manifestsName, appYamlName)
+		commitAndPush(sourceName, manifestsName, appYamlName)
+	} else {
+		fmt.Fprintf(shims.Stdout(), "Applying wego resources for application...")
+	}
 
 	fmt.Printf("Successfully added repository: %s.\n", params.Name)
 }
